Revert "Merge pull request #35 from UserOlena/schemasFolderFileStructure"
This reverts commit 7833b302a1e24ee54897a8d2041bd73380b2cf6e, reversing
changes made to a99f0e0f2ae6b66259885125036db971df412985.
