+++ server/package.json
@@ -11,14 +11,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "apollo-server": "^3.12.0",
     "apollo-server-express": "^3.6.2",
     "bcrypt": "^5.1.0",
     "dotenv": "^16.3.1",
     "express": "^4.17.2",
     "graphql": "^16.3.0",
     "jsonwebtoken": "^9.0.1",
-    "lodash": "^4.17.21",
     "mongoose": "^7.0.2"
   },
   "devDependencies": {
