--- server/schemas/lessonDefs.js
@@ -1,28 +1,0 @@
-const { gql } = require('apollo-server-express');
-const { 
-    Lesson, 
-  } = require('../models');
-
-// TODO: Complete Lesson typeDefs, queries, mutations
-export const typeDefs = gql`
-type Lesson {
-  _id: ID!
-  
-}
-
-type Query {
-
-}
-
-type Mutation {
-  
-}
-`;
-
-
-// TODO: Complete Resolvers for Lesson typeDefs
-  
-  export const resolvers = {
-    
-  };
-
