+++ server/schemas/typeDefs.js
@@ -1,0 +1,17 @@
+const { gql } = require('apollo-server-express');
+
+const typeDefs = gql`
+  type User {
+    _id: ID!
+    username: String
+    email: String
+    password: String
+  }
+
+  type Query {
+    user(_id: ID!): User!
+  }
+
+`;
+
+module.exports = typeDefs;
